+++ packages/sit-onyx/src/components/OnyxRadioGroup/OnyxRadioGroup.stories.ts
@@ -26,11 +26,11 @@ const EXAMPLE_OPTIONS: BaseSelectOption[] = [
 ];
 
 /**
- * This example shows a radio group with a headline.
+ * This example shows a radio group with a label.
  */
 export const Default = {
   args: {
-    headline: "Headline",
+    label: "Group label",
     options: EXAMPLE_OPTIONS,
     modelValue: EXAMPLE_OPTIONS[0].value,
   },
@@ -41,7 +41,7 @@ export const Default = {
  */
 export const Disabled = {
   args: {
-    headline: "Headline",
+    label: "Group label",
     options: EXAMPLE_OPTIONS,
     modelValue: EXAMPLE_OPTIONS[0].value,
     disabled: true,
@@ -53,7 +53,7 @@ export const Disabled = {
  */
 export const Horizontal = {
   args: {
-    headline: "Headline",
+    label: "Group label",
     options: EXAMPLE_OPTIONS,
     modelValue: EXAMPLE_OPTIONS[0].value,
     direction: "horizontal",