+++ packages/sit-onyx/CHANGELOG.md
@@ -1,5 +1,14 @@
 # sit-onyx
 
+## 1.0.0-beta.27
+
+### Major Changes
+
+- 3163863: fix: require aria label for radio and checkbox group
+
+  Removed property `headline` from `OnyxRadioGroup` and `OnyxCheckbox` in favor of new required `label` property which is also used as aria label for screen readers.
+  If you want to visually hide the label, set the `hideLabel` property.
+
 ## 1.0.0-beta.26
 
 ### Patch Changes